+++ templates/public/components/cart/cart-header.php
@@ -8,8 +8,8 @@
  */
 ?>
 
-<header class="bc-cart-header">
-	<div class="bc-cart-header__item"><?php esc_html_e( 'Item', 'bigcommerce' ); ?></div>
-	<div class="bc-cart-header__qty"><?php esc_html_e( 'Qty', 'bigcommerce' ); ?></div>
-	<div class="bc-cart-header__price"><?php esc_html_e( 'Price', 'bigcommerce' ); ?></div>
-</header>
+<div class="bc-cart-header" role="row">
+	<div class="bc-cart-header__item" role="columnheader"><?php esc_html_e( 'Item', 'bigcommerce' ); ?></div>
+	<div class="bc-cart-header__qty" role="columnheader"><?php esc_html_e( 'Qty', 'bigcommerce' ); ?></div>
+	<div class="bc-cart-header__price" role="columnheader"><?php esc_html_e( 'Price', 'bigcommerce' ); ?></div>
+</div>