+++ templates/public/components/cart/mini-cart.php
@@ -13,7 +13,7 @@
 
 ?>
 <!-- data-js="bc-cart" is required -->
-<section class="bc-cart bc-mini-cart" data-js="bc-cart" data-cart_id="<?php echo esc_attr( $cart['cart_id'] ); ?>">
+<section class="bc-cart bc-mini-cart" data-js="bc-cart" data-cart_id="<?php echo esc_attr( $cart['cart_id'] ); ?>" aria-label="<?php esc_attr_e( 'Mini Cart', 'bigcommerce' ); ?>" aria-live="polite" aria-relevant="additions removals">
 	<?php
 	echo $header;
 	echo $items;