+++ templates/public/components/currency-switcher-form.php
@@ -7,7 +7,7 @@
  */
 ?>
 
-<section class="bc-currency-switcher">
+<section class="bc-currency-switcher" aria-label="<?php esc_attr_e( 'Currency Switcher', 'bigcommerce' ); ?>">
 	<!-- data-js="bc-dynamic-fields" is required -->
 	<form class="bc-form" data-js="bc-dynamic-fields" method="post">
 		<?php wp_nonce_field( 'switch-currency' ); ?>
@@ -15,7 +15,7 @@
 
 		<div class="bc-form__control">
 			<label for="bc-currency-code"><?php esc_html_e( 'Currency Switcher', 'bigcommerce' ); ?></label>
-			<select name="bc-currency-code">
+			<select name="bc-currency-code" id="bc-currency-code">
 				<?php foreach ( $enabled_currencies as $currency ): ?>
 					<option value="<?php echo esc_html( $currency['currency_code'] ); ?>" <?php selected( $selected_currency, $currency['currency_code'] ); ?>><?php printf('%s (%s)', esc_html( $currency['name'] ), esc_html( $currency['token'] ) ); ?></option>
 				<?php endforeach; ?>
@@ -23,7 +23,7 @@
 		</div>
 
 		<div class="bc-form__actions">
-			<button class="bc-btn bc-btn--small bc-btn--form-submit" aria-label="<?php esc_html_e( 'Change Your Currency', 'bigcommerce' ); ?>" type="submit"><?php esc_html_e( 'Apply', 'bigcommerce' ); ?></button>
+			<button class="bc-btn bc-btn--small bc-btn--form-submit" aria-label="<?php esc_attr_e( 'Change Your Currency', 'bigcommerce' ); ?>" type="submit"><?php esc_html_e( 'Apply', 'bigcommerce' ); ?></button>
 		</div>
 	</form>
 </section>