+++ templates/public/components/option-types/option-swatch.php
@@ -11,9 +11,9 @@
 
 ?>
 <!-- class="bc-product-form__control bc-product-form__control--swatch" is required -->
-<div id="option-<?php echo esc_attr( $id ); ?>" class="bc-product-form__control bc-product-form__control--swatch">
+<fieldset id="option-<?php echo esc_attr( $id ); ?>" class="bc-product-form__control bc-product-form__control--swatch"<?php if ( $required ) { echo ' aria-required="true"'; } ?>>
 
-	<span class="bc-form__label bc-product-form__option-label <?php if ( $required ) { echo esc_attr( 'bc-form-control-required' ); } ?>"><?php echo esc_html( $label ); ?></span>
+	<legend class="bc-form__label bc-product-form__option-label <?php if ( $required ) { echo esc_attr( 'bc-form-control-required' ); } ?>"><?php echo esc_html( $label ); ?></legend>
 
 	<!-- data-js="product-form-option" and data-field="product-form-option-radio" are required -->
 	<div class="bc-product-form__option-variants bc-product-form__option-variants--inline" data-js="product-form-option" data-field="product-form-option-radio">
@@ -49,4 +49,4 @@
 		<?php } ?>
 	</div>
 
-</div>
+</fieldset>