+++ templates/public/components/reviews/product-rating.php
@@ -12,9 +12,13 @@ use BigCommerce\Post_Types\Product\Product;
 
 ?>
 <div class="bc-single-product__ratings">
-	<div class="bc-single-product__rating">
+	<div
+		class="bc-single-product__rating"
+		role="img"
+		aria-label="<?php echo esc_attr( sprintf( _n( 'Rated %s out of 5 stars, based on %d review', 'Rated %s out of 5 stars, based on %d reviews', $review_count, 'bigcommerce' ), number_format_i18n( $stars, 1 ), $review_count ) ); ?>"
+	>
 		<div class="bc-single-product__rating--mask" style="width: <?php echo (int) $percentage; ?>%">
-			<div class="bc-single-product__rating--top">
+			<div class="bc-single-product__rating--top" aria-hidden="true">
 				<span class="bc-rating-star"></span>
 				<span class="bc-rating-star"></span>
 				<span class="bc-rating-star"></span>
@@ -22,7 +26,7 @@ use BigCommerce\Post_Types\Product\Product;
 				<span class="bc-rating-star"></span>
 			</div>
 		</div>
-		<div class="bc-single-product__rating--bottom">
+		<div class="bc-single-product__rating--bottom" aria-hidden="true">
 			<span class="bc-rating-star"></span>
 			<span class="bc-rating-star"></span>
 			<span class="bc-rating-star"></span>